+++ decidim-bulletin_board-js/src/index.js
@@ -2,9 +2,9 @@ import { Client } from "./client/client";
 import { Election } from "./election/election";
 import { Trustee } from "./trustee/trustee";
 import { Voter } from "./voter/voter";
-
 import { MessageIdentifier } from "./client/message-identifier";
 import { MESSAGE_RECEIVED, MESSAGE_PROCESSED } from "./trustee/event_manager";
+import { KeyCeremonyComponent } from "./key-ceremony/key-ceremony.component";
 
 export {
   Client,
@@ -14,4 +14,5 @@ export {
   MessageIdentifier,
   MESSAGE_PROCESSED,
   MESSAGE_RECEIVED,
+  KeyCeremonyComponent,
 };